+++ filesfolders/plugins.py
@@ -55,9 +55,6 @@ class ProjectAppPlugin(ProjectAppPluginPoint):
     #: Search results template
     search_template = 'filesfolders/_search_results.html'
 
-    #: App card title for the main search page
-    search_title = 'Small Files, Folders and Links'
-
     #: App card template for the project details page
     details_template = 'filesfolders/_details_card.html'