+++ CHANGELOG.rst
@@ -6,7 +6,7 @@ Changelog for the SODAR Web UI. Loosely follows the
 guidelines.
 
 
-v0.3.1 (2018-08-XX)
+v0.3.1 (2018-08-24)
 ===================
 
 Added